Add a new MRO reading and sorting class to tmpl
The new class recursivley reads all other inherits for a class
form json files.
The class sorts all imports for all parent classes and then
return a properly sorted list for the imports of the
class being processed.

Class alos is built as a context manager. the class loads and
saves any cached results for 30 minutes by default.

With this caching a lot of repetivate json reads and proccssing
are avoided.