Const and Enum pyi not generate two files.
Const and Enum templates do double duty.
pyi class file is create in sun/_pyi/ subfolders for each
const and enum.

A secondary module (__init__.pyi) file is genereated in sun/ subfolder
For const this means a module containing const names only.
For enum this means a module with enum names only.

These kinds of imports are necessary to allow const and enum values
to be imported as single items.